Split out main image and content images, ready to treat them differently
