+++ packages/frontend/web/components/elements/MainImageComponent.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+import { ImageComponent } from '@frontend/web/components/elements/ImageComponent';
+
+export const MainImageComponent: React.FC<{
+    element: ImageBlockElement;
+    pillar: Pillar;
+}> = ({ element, pillar }) => {
+    return <ImageComponent element={element} pillar={pillar} />;
+};
